# Compiler: don't use init check for class vars initialized before read