--- a/src/compiler/crystal/codegen/class_var.cr
+++ b/src/compiler/crystal/codegen/class_var.cr
@@ -86,7 +86,9 @@ class Crystal::CodeGenVisitor
 
     # For unsafe class var we just initialize them without
     # using a flag to know if they were initialized
-    if class_var.uninitialized? || !init_func
+    if class_var.uninitialized? || !init_func || !class_var.read?
+      class_var.no_init_flag = true
+
       global = declare_class_var(class_var)
       global = ensure_class_var_in_this_module(global, class_var)
       if init_func
@@ -97,7 +99,7 @@ class Crystal::CodeGenVisitor
 
     global, initialized_flag = declare_class_var_and_initialized_flag_in_this_module(class_var)
 
-    lazy_initialize_class_var(initializer.node, init_func, global, initialized_flag)
+    lazy_initialize_class_var(initializer.node, init_func.not_nil!, global, initialized_flag)
   end
 
   def lazy_initialize_class_var(node, init_func, global, initialized_flag)
@@ -175,6 +177,8 @@ class Crystal::CodeGenVisitor
   end
 
   def read_class_var_ptr(class_var : MetaTypeVar)
+    class_var.read = true
+
     owner = class_var.owner
     case owner
     when VirtualType
@@ -184,7 +188,7 @@ class Crystal::CodeGenVisitor
     end
 
     initializer = class_var.initializer
-    if !initializer || class_var.uninitialized?
+    if !initializer || class_var.uninitialized? || class_var.no_init_flag?
       # Read directly without init flag, but make sure to declare the global in this module too
       return get_class_var_global(class_var)
     end