Similar to #9801 but for class variables.

# Benchmarks

Here's one benchmark to show how things improve, and more benchmarks to show that the standard library already benefits from this change.

## Generic benchmark for accessing a class variable

Run by passing `4` as an argument:

```crystal
require "benchmark"

class Foo
  class_property range = 3..7
end

a = 1
v = ARGV[0].to_i

Benchmark.ips do |x|
  x.report("class var") do
    case v
    when Foo.range
      a &+= 1
    end
  end
  x.report("inline") do
    case v
    when 3..7
      a &+= 1
    end
  end
  x.report("manual") do
    if 3 <= v <= 7
      a &+= 1
    end
  end
end

puts a
```

Before:

```
class var 299.07M (  3.34ns) (± 7.23%)  0.0B/op   2.74× slower
   inline 819.16M (  1.22ns) (± 6.16%)  0.0B/op        fastest
   manual 808.29M (  1.24ns) (± 6.69%)  0.0B/op   1.01× slower
```

After:

```
class var 670.64M (  1.49ns) (± 5.67%)  0.0B/op   1.19× slower
   inline 672.62M (  1.49ns) (± 7.11%)  0.0B/op   1.18× slower
   manual 796.03M (  1.26ns) (± 7.52%)  0.0B/op        fastest
```


## spawn

All fibers are stored in a stack_pool class variable.

```crystal
require "benchmark"

Benchmark.ips do |x|
  x.report("spawn") do
    spawn do
    end
  end
end
```

Before:

```
spawn 163.08k (  6.13µs) (±42.48%)  212B/op  fastest
```

After:

```
spawn 208.27k (  4.80µs) (±31.17%)  207B/op  fastest
```

## caller

This depends on a `skip` class variable to check for files that are skipped when building the caller stack.

```crystal
require "benchmark"

Benchmark.ips do |x|
  x.report("caller") do
    caller
  end
end
```

Before:

```
caller 152.18k (  6.57µs) (± 6.94%)  4.03kB/op  fastest
```

After:

```
caller 170.98k (  5.85µs) (± 3.64%)  2.71kB/op  fastest
```

## Array#to_s

This depends on a class variable to check for recursive execution.

```crystal
require "benchmark"

a = [1, 2, 3]
null = File.open(File::NULL, "w")

Benchmark.ips do |x|
  x.report("Array#to_s") do
    a.to_s(null)
  end
end
```

Before:

```
Array#to_s  12.84M ( 77.91ns) (± 2.26%)  0.0B/op  fastest
```

After:

```
Array#to_s  13.07M ( 76.52ns) (± 2.27%)  0.0B/op  fastest
```