@mattrberry 2020-11-30
@asterite I've recorded the performance of [CryBoy](https://github.com/mattrberry/CryBoy) on both 0.35.1 and this branch. For reference, the test Game Boy ROM I'm running is [pocket.gb](http://www.pouet.net/prod.php?which=65997). Unfortunately, between this change and #9801, performance seems to have only improved marginally. Still a useful improvement for other cases I think, but just figured I'd call out my specific example here!

`0.35.1`:
![image](https://user-images.githubusercontent.com/22065329/100560113-9e140880-3269-11eb-9f7e-a58750ecf160.png)
`opt/class-var-no-init-check`:
![image](https://user-images.githubusercontent.com/22065329/100560045-67d68900-3269-11eb-8682-b225ba39bb5c.png)