@mattrberry 2020-11-30
I'd be curious what the result would be if the `no_init_flag?` was made to be the first check in the `||` chain to short-circuit, for both class vars and constants. I don't have the time right now, but I'll try to play around with it tomorrow or the next day to see if that makes a noticeable difference.

Are the checks as they exist now in that order for a specific reason, or is that just the order they were added?