@mattrberry 2020-11-30
Actually, this just may be my ignorance regarding the implementation, but if constants can be assigned at compile-time (number / math expression / etc), is there a reason it's not just inlined everywhere? Why does the constant need to be looked up at a later point?