@straight-shoota 2020-11-30
A constant that gets assigned a primitive isn't much of a deal. But constants can have more complex values and depend on each other and on the runtime environment.