@asterite 2020-11-30
> I'd be curious what the result would be if the no_init_flag? was made to be the first check in the || chain to short-circuit, for both class vars and constants.

That code is run in the compiler when generating a program's code. It's not a check that's done at runtime in the generated program. So there will be no difference (well, maybe compilation will be faster but I doubt it because there are much more heavy things done by the compiler).

> Actually, this just may be my ignorance regarding the implementation, but if constants can be assigned at compile-time (number / math expression / etc), is there a reason it's not just inlined everywhere? Why does the constant need to be looked up at a later point?

I noticed this actually happens in your code, in `opcodes.cr`. You have an array `UNPREFIXED` of procs, and one of the procs accesses `PREFIXED` which is declared **after** `UNPREFIXED`. What if we try to initialize `UNPREFIXED` and we actually need to execute one of those procs (the compiler can't reliably know this will happen or not). Well, `PREFIXED` won't exist yet so it will crash. So, because `PREFIXED` is accessed before it's declared, the compiler decided that it must be lazily initialized.

One way to solve this, and it might make your program faster, is to move the definition of `UNPREFIXED` **after** `PREFIXED`.

Another way to generally "solve" this in the language would be to yell if you try to "use" a constant before you declare it. But note that in your example you don't actually use it, and the compiler has no way to know this. So, the compiler is more permissive at the cost of making things (a bit) slower. But maybe all of this should be more explicit and we really need to give an error here.

Another such lazily initialized constant is `FETCHER_ORDER`, though I didn't trace why yet. I think it's because some memory.tick happens inside one of those procs I mentioned before, and that ends up using that constant **before** it was declared. This is fixed by adding `require "./fifo_ppu" inside `memory.cr`, then `require "./ppu"` inside `fifo_ppu.cr`. In general I recommend making these `require` dependencies super explicit, that way you can guarantee the order of the initialization of constants.